+++ cloudpickle/compat.py
@@ -1,7 +1,9 @@
 import sys
+import platform
 
+PYPY = platform.python_implementation() == "PyPy"
 
-if sys.version_info < (3, 8):
+if sys.version_info < (3, 8) or PYPY:
     try:
         import pickle5 as pickle  # noqa: F401
         from pickle5 import Pickler  # noqa: F401